floating points

+++ inss.py
@@ -1,14 +1,15 @@
 import csv
+from decimal import Decimal
 
 class InssRate:
     def __init__(self, min=0, max=0, rate=0):
-        self.min=min
-        self.max = max
-        self.rate =rate
+        self.min= Decimal(min)
+        self.max = Decimal(max)
+        self.rate =Decimal(rate)
     
     def calculate(self, value):
         base = value if value< self.max else self.max
-        value = base * self.rate / 100
+        value = Decimal(base * self.rate / 100)
         return round(value,2)
 
 def load_inss(csv_file_path='inss.csv'):
@@ -16,7 +17,7 @@ def load_inss(csv_file_path='inss.csv'):
     with open(csv_file_path) as csvfile:
         reader = csv.DictReader(csvfile)
         for row in reader:
-            o = InssRate(float(row['min']), float(row['max']), float(row['rate']))
+            o = InssRate(Decimal(row['min']), Decimal(row['max']), Decimal(row['rate']))
             insss.append(o)
     return insss
