+++ irrf.py
@@ -1,18 +1,19 @@
 import csv
+from decimal import Decimal
 
 class IrrfRate:
     def __init__(self, min=0, max=0, rate=0, deduction=0, dependents_reduction=0):
-        self.min=min
-        self.max = max
-        self.rate =rate
-        self.deduction = deduction
-        self.dependents_reduction = dependents_reduction
+        self.min= Decimal(min)
+        self.max = Decimal(max)
+        self.rate =Decimal(rate)
+        self.deduction = Decimal(deduction)
+        self.dependents_reduction = Decimal(dependents_reduction)
     
     def calculate(self, base_value, dependents=0):
         value = 0
         if self.min <= base_value <= self.max:
-            base = base_value if dependents==0 else base_value - (dependents * self.dependents_reduction)
-            rate = base * self.rate / 100
+            base = Decimal(base_value) if dependents==0 else Decimal(base_value) - (dependents * self.dependents_reduction)
+            rate = Decimal(base * self.rate / 100)
             value = rate - self.deduction
         return round(value,2)
 
@@ -28,7 +29,7 @@ def load_irrf(csv_file_path='irrf.csv', dependents_reduction=0):
     with open(csv_file_path) as csvfile:
         reader = csv.DictReader(csvfile)
         for row in reader:
-            o = IrrfRate(float(row['min']), float(row['max']), float(row['rate']), float(row['deduction']), dependents_reduction)
+            o = IrrfRate(Decimal(row['min']), Decimal(row['max']), Decimal(row['rate']), Decimal(row['deduction']), dependents_reduction)
             irrfs.append(o)
     return irrfs
